{{ cookiecutter.package_name }}-project/pyproject.toml
[tool.poetry]
name = "{{ cookiecutter.package_name }}"
version = "{{ cookiecutter.semantic_version }}"
description = "Sample AWS Human-in-the-Loop Python project with production ready setup"
authors = ["{{ cookiecutter.author_name }} <husanhe@gmail.com>"]


[tool.poetry.dependencies]
python = "3.8.*"
aws_lambda_event = "2.0.1"
aws_console_url = "0.5.1"
botocore = "1.27.96"
boto3 = "1.24.96"
boto_session_manager = "1.3.2"
boltons = "21.0.0"
fixa = "0.3.2"
config_patterns = "0.3.1"
chalice = "1.27.3"
attrs = "21.4.0"
attrs_mate = "1.0.2"
jmespath = "1.0.1"
pathlib_mate = " 1.2.1"
s3pathlib = "1.3.1"
smart-open = "6.2.0"
superjson = "1.0.2"
pysecret = "2.2.2"
pynamodb = "5.3.4"
pynamodb-mate = "5.3.4.8"
aws_a2i = "0.2.1"
light_emoji = "{{ cookiecutter.semantic_version }}"


[tool.poetry.dev-dependencies]
# Don't put anything here, explicit is better than implicit


[tool.poetry.group.dev]
optional = true


[tool.poetry.group.dev.dependencies]
twine = "3.8.0"
wheel = "0.37.1"
rich = "12.5.1"
fire = "0.5.0"
cottonformation = "1.0.1"
aws_cloudformation = "1.2.1"
python-box = "6.1.0"
python-liquid = "1.6.1"


[tool.poetry.group.doc]
optional = true


[tool.poetry.group.doc.dependencies]
Sphinx = "4.3.0"
sphinx-jinja = "1.1.1"
rstobj = "1.2.1"
docfly = "1.0.2"
furo = "2021.8.31"
sphinx-inline-tabs = "2021.8.17b10"


[tool.poetry.group.test]
optional = true


[tool.poetry.group.test.dependencies]
pytest = "6.2.5"
pytest-cov = "2.12.1"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
